refactor: use custom link
